+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 20fedd5 - 2017-11-05: Fixed search issues for RARBG 
+- * 8459360 - 2017-11-05: Fixed search issues for RARBG 
+- * b3feaac - 2017-11-05: Fixed search issues for RARBG 
 - * c5fa018 - 2017-11-05: Fixed issue with messages being NoneType during post-processing 
 - * 31c2cbe - 2017-11-05: Fixed regex&#x27;s for Newpct provider and added search url for HD series 
 - * 1260ed4 - 2017-11-05: Fixed regex&#x27;s for Newpct provider and added search url for HD series